refactor: move builtin to a separate file
